Move awaitAsync to a new file
