+++ Tests/ChangeLogGeneratorTests/GeneratorSpec.swift
@@ -194,33 +194,3 @@ final class GeneratorSpec: QuickSpec {
         }
     }
 }
-
-struct NoResultError: Error {}
-
-func awaitAsync<T>(file: FileString = #file, line: UInt = #line, _ function: @escaping () async throws -> T) throws -> T {
-    func fetch(completion: @escaping (Result<T, Error>) -> Void) {
-        Task {
-            do {
-                completion(.success(try await function()))
-            }
-            catch {
-                completion(.failure(error))
-            }
-        }
-    }
-
-    var outputs: [Result<T, Error>] = []
-
-    waitUntil(timeout: .seconds(5)) { done in
-        fetch { result in
-            outputs.append(result)
-            done()
-        }
-    }
-
-    guard let output = outputs.first else {
-        throw NoResultError()
-    }
-
-    return try output.get()
-}
